+++ include/crprod.hpp
@@ -3,7 +3,7 @@
 
 class CRprod final:public CRobj {
     public: 
-        CRprod(int i, size_t l);
+        CRprod(long long i, size_t l);
         
         std::unique_ptr<CRobj> add(const CRobj& target) const override;
         std::unique_ptr<CRobj> mul(const CRobj& target) const override;
@@ -17,7 +17,7 @@ class CRprod final:public CRobj {
         std::unique_ptr<CRobj> cos() const override;
 
         void simplify() override;
-        void shift(size_t i) override final { 
+        void shift(long long i) override final { 
             if (index > i){
                 for (size_t j = 0; j < isanumber.size(); j++){ 
                     operands[isanumber[j]]->shift(i);
@@ -30,7 +30,7 @@ class CRprod final:public CRobj {
             }
         }
         void print_tree() const override;
-        std::string genCode(size_t parent, size_t index, int place,std::string indent) const override;
+        std::string genCode(size_t parent, long long index, long long place,std::string indent) const override;
